moves alignment to utils
