+++ src/utils/mod.rs
@@ -1,3 +1,5 @@
+pub mod alignment;
 pub mod data_ingest;
 pub mod data_processing;
+pub mod read_fastq;
 pub mod writing_outputs;
